+++ example/08_sky/08_sky.cpp
@@ -66,7 +66,7 @@ void ex_08_sky(Shell& app, Widget& parent, Dockbar& dockbar)
 	SceneViewer& viewer = ui::scene_viewer(parent);
 	ui::orbit_controller(viewer);
 
-	if(Widget* dock = ui::dockitem(dockbar, "Game", carray<size_t, 1>{ 1U }))
+	if(Widget* dock = ui::dockitem(dockbar, "Game", carray<uint16_t, 1>{ 1U }))
 		example_options(*dock, g_sky);
 
 	Gnode& groot = viewer.m_scene->begin();
